+++ models/amendmentNumbering/PerMotionEnglish.php
@@ -1,7 +1,10 @@
 <?php
+
+declare(strict_types=1);
+
 namespace app\models\amendmentNumbering;
 
-use app\models\db\{Amendment, Motion};
+use app\models\db\{Amendment, IMotion};
 
 class PerMotionEnglish extends IAmendmentNumbering
 {
@@ -15,19 +18,22 @@ class PerMotionEnglish extends IAmendmentNumbering
         return 3;
     }
 
-    public function getAmendmentNumber(Amendment $amendment, Motion $motion): string
+    /**
+     * @param Amendment[] $otherAmendments
+     */
+    public function getAmendmentNumber(Amendment $amendment, IMotion $baseImotion, array $otherAmendments): string
     {
         $prefixes = [];
-        foreach ($motion->amendments as $amend) {
+        foreach ($otherAmendments as $amend) {
             $prefixes[] = $amend->titlePrefix;
         }
         $maxRev = static::getMaxTitlePrefixNumber($prefixes);
 
-        $str = trim($motion->titlePrefix);
+        $str = trim($baseImotion->titlePrefix);
         if ($str) {
             $str .= ' ';
         }
-        $str .= 'A' . ($maxRev + 1);
+        $str .= \Yii::t('amend', 'amendment_prefix') . ($maxRev + 1);
 
         return $str;
     }