+++ History.md
@@ -9,6 +9,7 @@
 - The export functions for motions in the admin list now allow to combine one or multiple motion types, with all motion types being selected by default.
 - The consultation-wide search now also searches for the name of proposers of motions / amendments.
 - The small-screen main menu now also has the full-text search.
+- Amendments are now prefixed with "AM", not "Ä". The prefix can be changed through the language-setting.
 - Amendments to Statute Amendments are now better supported:
   - On the home page, those amendments are now shown like regular amendments, not like stand-alone motions.
   - These amendments now show the base statute amendment and the changed version next to each other, for easier comparison.